+++ sdmx-dl-provider-ri/src/main/java/module-info.java
@@ -15,11 +15,12 @@ module sdmxdl.web.ri {
 
     provides sdmxdl.web.spi.SdmxWebDriver with
             internal.sdmxdl.ri.web.drivers.AbsDriver2,
+            internal.sdmxdl.ri.web.drivers.BbkDriver,
             internal.sdmxdl.ri.web.drivers.DotStatDriver2,
             internal.sdmxdl.ri.web.drivers.FileDriver,
+            internal.sdmxdl.ri.web.drivers.InseeDriver2,
             internal.sdmxdl.ri.web.drivers.NbbDriver2,
-            internal.sdmxdl.ri.web.drivers.Sdmx21Driver2,
-            internal.sdmxdl.ri.web.drivers.BbkDriver;
+            internal.sdmxdl.ri.web.drivers.Sdmx21Driver2;
 
     provides sdmxdl.file.spi.SdmxFileReader with
             internal.sdmxdl.ri.file.readers.XmlFileReader;