+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add support of codelist in API [#158](https://github.com/nbbrd/sdmx-dl/issues/158)
+- Add native driver for INSEE [#157](https://github.com/nbbrd/sdmx-dl/issues/157)
 
 ### Changed